uintmaxtoa

+++ ft_memcpy.c
@@ -6,7 +6,7 @@
 /*   By: pgerbaud <pgerbaud@student.42.fr>          +#+  +:+       +#+        */
 /*                                                +#+#+#+#+#+   +#+           */
 /*   Created: 2016/11/09 10:42:21 by pgerbaud          #+#    #+#             */
-/*   Updated: 2016/11/17 17:19:55 by pgerbaud         ###   ########.fr       */
+/*   Updated: 2017/12/08 17:45:49 by pgerbaud         ###   ########.fr       */
 /*                                                                            */
 /* ************************************************************************** */
 
@@ -14,10 +14,10 @@
 
 void	*ft_memcpy(void *restrict dst, const void *restrict src, size_t n)
 {
-	size_t	index;
+	int		index;
 
 	index = 0;
-	while (index < n)
+	while (index < (int)n)
 	{
 		*(char *)(dst + index) = *(char *)(src + index);
 		index++;
